Fix `TestHardLinksInArchive` to pass if fails on windows as task user
